Move barometer updates to the inertial ISR
Some combination on inconsistent update timing and the MPU's SPI bus usage
at non-harmonious intervals with the barometer is causing large deviations
in altitude in between readings.

By putting the relativly short barometer update function in the flight
stabilization ISR, the SPI bus usage interval is consistent and the
barometer is updated an even pace. This has removed most of the erronious
noise observed in the barometer when running the quadcopter sketch.

My attempts to recreate the exact behaviour in a simpler sketch
with long interrupts and other SPI readings have failed; but SPI bus usage
is noted as interfering with MS5611 ADC readings in its datasheet. What
exactly was happening remains a mystery. Limiting the quadcopter ISR
to just the IMU update causes the error to go away, but running the
barometer with nothing but `delay` equiped ISR's also fails to cause
the error.