fix for HMC5883L Clone
Fixed code so that for compasses that do not use a genuine HMC5883L the clone will function. If status check of register 0x01 succeeds continues as a true HMC5883L. If the register check fails switch to using clone after first testing if the values returned  by the clone are reasonable. The clone does not return a good value for read of register 0x01. The clone does not require a prompt to update register values. The clone is read after a write to 0x03 instead of two writes to 0x02 and 0x01 successively.