put checkGoodValues in safe() for ARM time check
Setup safe() function so that it now performs the compass "goodValuesCheck()" to check if compass is returning reasonable values. This safe() function is called when the quadcopter arms. The status function of HMC5883L now only sets the compass type. It always returns Sensor::OK. This transfers the check of reasonable compass values from the status function which was called from the Arduino "setup()" to the ARM step which is called whenever the arm sequence is started.