fix compiler crash, set compass check at ARM time
Split into .h and .cpp to fix compiler segementation fault at compile time. Set the status() function which is called when the board is powered on to only set which time of compass is used but not check compass functionality. status() now always returns Sensor::OK. The compass functionality check is now in checkGoodValues which is called by safe() in ardupilot.h. safe() is called when the quadcopter is armed. Compass is now checked for good values when the quadcopter is armed instead of when it is powered.