fix not arming raising max on compass
Not checking for max raw compass values anymore unless over 9999 (max should really be less than 1000). Was preventing drone from arming. Yaw still appears to be stabile in flight. Might be due to instantaneous check instead of average. These values appear to change when drone is not connected to computer. Was arming when connected to computer via usb but not otherwise.