Compass clone fix (#1)
* fix for HMC5883L Clone

Fixed code so that for compasses that do not use a genuine HMC5883L the clone will function. If status check of register 0x01 succeeds continues as a true HMC5883L. If the register check fails switch to using clone after first testing if the values returned  by the clone are reasonable. The clone does not return a good value for read of register 0x01. The clone does not require a prompt to update register values. The clone is read after a write to 0x03 instead of two writes to 0x02 and 0x01 successively.

* fix compiler warning for pointer type

fix compiler warning by changing false type to 0 instead so that pointer returned is of the correct type.

* fix min battery voltage to work with new quad

Reduce lowest acceptable battery voltage to 10.5 volts so that the current quadcopter model will fly without further modification of the code or a reset from the dashboard.

* remove extra include sensor.h

Since sensor.h is already included in "input/InertialManager.h" it does not need to be included again

* remove sensor.h from the includes

sensor.h is included by "input/InertialManager.h". therefore it does  not need to be included again. The code is commented instead of deleted, since this file can be used as a reference to find which directories other files are in.

* split sensor.h into .h & .cpp fix multi def error 

multiple define error for Sensor::OK occurs when the line in the .cpp file is include in the header instead.

* put checkGoodValues in safe() for ARM time check

Setup safe() function so that it now performs the compass "goodValuesCheck()" to check if compass is returning reasonable values. This safe() function is called when the quadcopter arms. The status function of HMC5883L now only sets the compass type. It always returns Sensor::OK. This transfers the check of reasonable compass values from the status function which was called from the Arduino "setup()" to the ARM step which is called whenever the arm sequence is started.

* fix compiler crash, set compass check at ARM time

Split into .h and .cpp to fix compiler segementation fault at compile time. Set the status() function which is called when the board is powered on to only set which time of compass is used but not check compass functionality. status() now always returns Sensor::OK. The compass functionality check is now in checkGoodValues which is called by safe() in ardupilot.h. safe() is called when the quadcopter is armed. Compass is now checked for good values when the quadcopter is armed instead of when it is powered.

* fix for TGY-i6S transmitter

Switched pitch and roll channels so that if channels are plugged in ch1->ch4 on transmitter to input pin 1->4 on flight controller pitch and roll control the proper channels. This involves exchanging a 1<->0 at top of file.

* added explicit mode set for Clone

Added lines to set the mode register 0x02 to continuous mode 0x00 if compass fails status  check (check of register 0x09) indicating that it is a clone instead of a true HMC5883L. The clone seems to only work in continuous mode.

* creating directory for quadcopter_gimbal

* code for quadcopter on gimbal

Code is setup for use with quadcopter mounted on gimbal. Throttle is limited to below 50% and the assisted mode (altitude hold) is disabled.

* use continous mode for both HMC5883L & clone

Change to using continuous mode for the original HMC5883L as well as the clone. From flight test this is steadier. Compass is now sent into continuous mode: register 0x02 set to 0x00 in the "begin()" function.  "checkGoodValues()" now checks the actual values instead of the 0x09 register for both compasses. The clone flag "isTruHMC5883L" is still set by "status()" but it is no longer used every read call. The compass should be averaging over 8 measurements and running at 15Hz as set by 0x70 being sent into 0x00.

* quick fix for not arming

Drone is arming only when connected to the computer when max expected magnitude is at 1500 instead of 2500. This is a temporary fix.

* fix not arming raising max on compass

Not checking for max raw compass values anymore unless over 9999 (max should really be less than 1000). Was preventing drone from arming. Yaw still appears to be stabile in flight. Might be due to instantaneous check instead of average. These values appear to change when drone is not connected to computer. Was arming when connected to computer via usb but not otherwise.