Fix for missing groundSettings.xml entry
The settings parsing script/regex appears to skip the first ground settings entry in the .ino file.

Added a dummy entry to be thrown away instead of the first real settings item in the list. A better future fix would be to investigate the regular expression (which is not commented or easily readable of course...) found in the settingsDescriptionParser.py file of the dashboard source files.

Also added a missing index assignment in the setupSettings function.